+++ examples/basic-project/tsconfig.json
@@ -27,6 +27,6 @@
       "ice": [".ice"]
     }
   },
-  "include": ["src", ".ice"],
+  "include": ["src", ".ice", "ice.config.*"],
   "exclude": ["node_modules", "build", "public"]
 }
