+++ packages/build-webpack-config/package.json
@@ -32,7 +32,7 @@
   },
   "devDependencies": {
     "@ice/types": "^1.0.0",
-    "build-scripts": "^2.0.0-12",
+    "build-scripts": "^2.0.0-13",
     "webpack": "^5.69.1",
     "webpack-dev-server": "^4.7.4"
   }
