+++ packages/ice/package.json
@@ -26,7 +26,7 @@
     "@ice/route-manifest": "^1.0.0",
     "@ice/runtime": "^1.0.0",
     "address": "^1.1.2",
-    "build-scripts": "^2.0.0-12",
+    "build-scripts": "^2.0.0-13",
     "chalk": "^4.0.0",
     "commander": "^9.0.0",
     "consola": "^2.15.3",
