+++ packages/types/src/index.ts
@@ -1,9 +1,11 @@
 import type { Config } from './config.js';
 import type { Plugin } from './plugin.js';
+import type { UserConfig } from './userConfig.js';
 
 export {
   Config,
   Plugin,
+  UserConfig,
 };
 
 export * from './runtime.js';
