+++ packages/types/src/userConfig.ts
@@ -3,12 +3,12 @@ import type { Config } from './config';
 export interface UserConfig {
   alias?: Record<string, string | false>;
   define?: Record<string, string | boolean>;
-  devPublicPath: string;
-  publicPath: string;
-  hash: boolean | string;
-  externals: Config['externals'];
-  outputDir: string;
-  proxy: Config['proxy'];
-  filename: string;
-  webpack: Config['configureWebpack'][0];
+  devPublicPath?: string;
+  publicPath?: string;
+  hash?: boolean | string;
+  externals?: Config['externals'];
+  outputDir?: string;
+  proxy?: Config['proxy'];
+  filename?: string;
+  webpack?: Config['configureWebpack'][0];
 }
